perf(webui): lazy-load the connect-link flow off the initial /chat chunk
The connect-link validation added in the previous commit pushed the eager
/chat closure to 222.2 KB gzip, over its 222.0 KB budget: resolving the
`connect` param pulled the extensions API client and surface schema into
the initial route.

Move the resolver, the install/oauth-start sequence, and the flow watcher
into `pages/chat/lib/connect-link-flow.ts`, loaded through a dynamic
`import()` only when a `connect` param is actually present. The hook keeps
the param detection, the URL strip, and the card state, all of which are
tiny. The click handler still opens its placeholder popup synchronously
before awaiting the module, so the user activation is not burned.

Measured with `vite build` + `scripts/check-bundle-budgets.ts`: 222.2 KB
over budget -> 221.4 KB, which is also 0.1 KB below the 221.5 KB the same
check measures on the pre-PR tree.

No behavior change; all 14 useConnectLinkLanding tests still pass (the
suite now imports the flow module statically so the dynamic import
resolves on a microtask).
